Rename files to match class names
