+++ gpytorch/priors/__init__.py
@@ -3,9 +3,9 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from .gamma import GammaPrior
-from .multivariate_normal import MultivariateNormalPrior
-from .normal import NormalPrior
-from .smoothed_box import SmoothedBoxPrior
+from .gamma_prior import GammaPrior
+from .multivariate_normal_prior import MultivariateNormalPrior
+from .normal_prior import NormalPrior
+from .smoothed_box_prior import SmoothedBoxPrior
 
 __all__ = [GammaPrior, MultivariateNormalPrior, NormalPrior, SmoothedBoxPrior]
