+++ gpytorch/priors/gamma_prior.py
