+++ gpytorch/priors/smoothed_box_prior.py
